keymaker bandicam pass 123.zip
DOWNLOAD AND INSTALL KEYMAKER FOR BANDICAM
Bandicam is a recognized leader among video capture applications. You can use it to record an image from a PC screen or a webcam. It is natural that such useful software is paid. Only the version with restrictions is freely available. But you can get access to all the functionality using KeyMaker. How exactly, we will tell you further. But first we note that the full version can be downloaded by bandicam from the main page.

DOWNLOAD AND INSTALL KEYMAKER FOR BANDICAM
Bandicam is a recognized leader among video capture applications. You can use it to record an image from a PC screen or a webcam. It is natural that such useful software is paid. Only the version with restrictions is freely available. But you can get access to all the functionality using KeyMaker. How exactly, we will tell you further. But first we note that the full version can be downloaded by bandicam from the main page.

Download KayMaker ↓ How to close access to the Bandicam network ↓

STEP 1-DOWNLOAD THE CORRECT VERSION OF BANDICAM
KeyMaker was tested with the latest version of Bandicam, tested Bandicam download from the main page of the site.

Before starting the bandicam activator, we recommend stopping the antivirus. Otherwise, the antivirus program may mistake KeyMaker for a virus and delete it. As a result, you will have to download it again or search for it in quarantine and restore it. (At your own risk)

STEP 2-DOWNLOAD AND UNPACK KEYMAKER
Download the archive with the application and unpack it using the archiver or file manager. At the same time, Bandicam must be installed in advance. If you do not do this, you will not be able to activate the full version of the program.
Password: 123

KeyMaker will inform you that the program must be installed first.

STEP 3-LAUNCH THE ACTIVATOR
You need to act on behalf of the administrator.

before launching, we must close Bandicam, if it was open.

STEP 4-REGISTER BANDICAM
The only thing you need to do to unlock the full version of the program is to enter an email address if the window for it is active and there is no reminder to install the application. At the same time, it is not necessary to enter your real e-mail. You can take it, but with changes, or even an arbitrary combination of characters that resembles an email address.

We start the registration of the application by clicking on the inscription "Register application".

If everything was successful, a window will appear informing you about it:

Bandicam Registred Succesfully!

STEP 5-CLOSE BANDICAM ACCESS TO THE NETWORK
Your Bandicam will be activated, but this will be until you manually update the license information or the program will do it itself after some time, in this case the license will fly off, you can use the activator again, but if you do not want this to happen again, you need to block access to the Internet for the program so that it does not check activation.

There are two ways to do this:

* add the application to the firewall filter;
* edit the hosts file.

We will not consider the first option, since each of the numerous firewalls has its own principle of operation. Our example shows the process of editing the hosts file. Regardless of the version of KeyMaker, the algorithm of actions is similar.

Open host in notepad or another text editor. You can find it at – C:\Windows\System32\drivers\etc.

Add two lines to the file:

127.0.0.1 bandicam.com
127.0.0.1 www.bandicam.com

STEP 6-CHECK THE RESULT
It remains only to see whether it was possible to register Bandicam. We launch the application and look for the inscription "Not registered"in the header. If it's not there, it's fine. The license information is available in the "About the program" menu section.